Initialize the ApplicationBeans id in PostConstruct, not constructor
Move the initialization of the the id field in ApplicationBeans from the
constructor to a PostConstruct method.

The test that uses this class verifies that only one instance of the
ApplicationScoped bean is created.

However, because this is a normal scoped bean, the container may need to
create client proxies for this class and when it constructs one of
these, it will call the zero-arg constructor.

Therefore, although only one instance of the bean will be created, its
constructor may be called more than once.

Setting up the id in a PostConstruct method avoids this issue because
the PostConstruct method won't be called when creating a client proxy.