Set imagesdir so that images appear on github
The maven plugin defaults imagesdir to "images", but github doesn't do
that when it renders the preview for asciidoc files.

This change allows images to show up correctly in github previews, as
well as in the built documents.

Github also doesn't handle includes, so you can only view each file
individually so we need to specify imagesdir in every file.

Signed-off-by: Andrew Rouse <anrouse@uk.ibm.com>