Fix two emitter tests
The BeanMissingChannel should fail at startup as there is no `downstream` channel for the emitter. It's a definition exception (equivalent to injecting a missing bean).

The ChannelInjectionDifferentFlavourSameChannelTest is invalid as a single upstream cannot, by default, emit values to multiple (6 in the case of the tests) downstream. Implementations could support such behavior, but the TCK should not assume that this mechanism exists or is the default behavior.

Signed-off-by: Clement Escoffier <clement.escoffier@gmail.com>