Reorder waits and assertions to provide better result assurance
Wait for the failure on the stream to occur instead of the exception on the emitter - this ensures that at least the first message will be sucessfully processed and that a failure did occur before assertions are checked.

In the case where the failure may occur sufficiently late in the test execution such that there is a failure, but not yet an exception. In this case emit one more message and wait for the exception before checking

emitThree was unused and would close close the stream via `.complete()` - repurpose for being able to send one message and not close the stream if successful.